Move the setting of msgs_good and msgs_bad into msgcounts.c, where minima and maxima can be enforced.  Convert msgs_good and msgs_bad from u_int32_t to double.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@4217 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ src/globals.h
@@ -54,8 +54,8 @@ extern	const char *const version;
 extern	const char *const system_config_file;
 
 /* for msgcounts.c */
-extern	u_int32_t	 msgs_good;
-extern	u_int32_t	 msgs_bad;
+extern	double		 msgs_good;
+extern	double		 msgs_bad;
 
 /* for passthrough.c */
 extern size_t msg_register_size;
